# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Filkx API",
        "description": "API documentation for the Tech Filkx e-commerce platform",
        "contact": {
            "email": "admin@filkx.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://dev.api.tech.filkx.com",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/v1/checkout": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Place an order from the current cart",
                "operationId": "d688169a8d59187bed63f5593a20f961",
                "parameters": [
                    {
                        "name": "X-Cart-Session-ID",
                        "in": "header",
                        "description": "Guest cart session identifier. Omit when authenticated - the customer's own cart is used instead.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customerName",
                                    "customerPhone",
                                    "customerEmail",
                                    "shippingAddress",
                                    "deliveryMethod",
                                    "paymentMethod"
                                ],
                                "properties": {
                                    "customerName": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "customerPhone": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "customerEmail": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "shippingCountry": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 100
                                    },
                                    "shippingCity": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 100
                                    },
                                    "shippingAddress": {
                                        "type": "string",
                                        "maxLength": 500
                                    },
                                    "deliveryMethod": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "paymentMethod": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "sessionId": {
                                        "description": "Alternative to the X-Cart-Session-ID header.",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "couponCode": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created from the cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CheckoutOrderResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Cart is empty or a line item/coupon failed validation (out of stock, inactive product, invalid coupon, ...)"
                    }
                }
            }
        },
        "/api/v1/checkout/quick": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Place a quick single-item order (no cart, no coupon, fixed Kyiv/Nova Poshta shipping placeholder)",
                "operationId": "5fe7fbd8e465739dea2f7b1742f8920d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customerName",
                                    "customerPhone",
                                    "variantId"
                                ],
                                "properties": {
                                    "customerName": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "customerPhone": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "variantId": {
                                        "type": "integer"
                                    },
                                    "paymentMethod": {
                                        "type": "string",
                                        "default": "cod",
                                        "enum": [
                                            "cod",
                                            "card"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quick order created for a single unit of the given variant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CheckoutOrderResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed, or the variant is unavailable/out of stock"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CheckoutOrderResource": {
                "title": "Checkout Order Resource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "orderNumber": {
                        "type": "string",
                        "example": "FKX-20260818-AB12CD"
                    },
                    "userId": {
                        "type": "integer",
                        "nullable": true
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "customerEmail": {
                        "type": "string",
                        "format": "email"
                    },
                    "customerPhone": {
                        "type": "string"
                    },
                    "shippingCountry": {
                        "type": "string",
                        "nullable": true
                    },
                    "shippingCity": {
                        "type": "string",
                        "nullable": true
                    },
                    "shippingAddress": {
                        "type": "string"
                    },
                    "deliveryMethod": {
                        "type": "string"
                    },
                    "paymentMethod": {
                        "type": "string"
                    },
                    "paymentStatus": {
                        "type": "string",
                        "example": "pending"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending_payment"
                    },
                    "totalPrice": {
                        "type": "number",
                        "format": "float"
                    },
                    "couponCode": {
                        "type": "string",
                        "nullable": true
                    },
                    "discountAmount": {
                        "type": "number",
                        "format": "float"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "orderId": {
                                    "type": "integer"
                                },
                                "variantId": {
                                    "type": "integer"
                                },
                                "productName": {
                                    "type": "string"
                                },
                                "sku": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "quantity": {
                                    "type": "integer"
                                },
                                "createdAt": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "updatedAt": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}